# crates/bunker-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LAN BUNKER API",
    "description": "The LAN BUNKER backend over SQLite: players and auth, tournaments and brackets, events and check-ins, and the cycles ledger",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/tournaments": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "list_tournaments",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Every tournament, drafts included",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_Tournament"
                }
              }
            }
          },
          "400": {
            "description": "A query parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin, or `PasswordChangeRequired`: a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "create_tournament",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTournament"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "A new draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tournament"
                }
              }
            }
          },
          "400": {
            "description": "The body is not valid JSON, or a parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin, or `PasswordChangeRequired`: a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "The body is too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "415": {
            "description": "The body is not `application/json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "A field is missing, unknown or out of range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/tournaments/{id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "get_tournament",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TournamentId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDetail"
                }
              }
            }
          },
          "400": {
            "description": "A path parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin, or `PasswordChangeRequired`: a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "operationId": "delete_tournament",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TournamentId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Gone, with its entrants and matches"
          },
          "400": {
            "description": "A path parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin, or `PasswordChangeRequired`: a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "operationId": "update_tournament",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TournamentId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TournamentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tournament"
                }
              }
            }
          },
          "400": {
            "description": "The body is not valid JSON, or a parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin, or `PasswordChangeRequired`: a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The tournament is concluded, or changed in the meantime",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "The body is too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "415": {
            "description": "The body is not `application/json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "A field is missing, unknown or out of range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/me/registrations": {
      "get": {
        "tags": [
          "tournaments"
        ],
        "operationId": "registrations",
        "responses": {
          "200": {
            "description": "The tournaments the caller entered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Registrations"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "`PasswordChangeRequired`: the caller must choose a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/tournaments": {
      "get": {
        "tags": [
          "tournaments"
        ],
        "operationId": "list_tournaments",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Newest event first. Drafts are hidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_Tournament"
                }
              }
            }
          },
          "400": {
            "description": "A query parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}": {
      "get": {
        "tags": [
          "tournaments"
        ],
        "operationId": "get_tournament",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TournamentId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDetail"
                }
              }
            }
          },
          "400": {
            "description": "A path parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Unknown, or still a draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}/registration": {
      "post": {
        "tags": [
          "tournaments"
        ],
        "operationId": "register",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TournamentId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The caller's entry. A second call answers the same one, with the new level",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entrant"
                }
              }
            }
          },
          "400": {
            "description": "The body is not valid JSON, or a parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "`PasswordChangeRequired`: the caller must choose a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Registration is not open, or the field is full",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "The body is too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "415": {
            "description": "The body is not `application/json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "The level is not 1 to 5",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "tournaments"
        ],
        "operationId": "retire",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TournamentId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Gone, or never in"
          },
          "400": {
            "description": "A path parameter is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "`PasswordChangeRequired`: the caller must choose a new password first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Registration is not open",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "The one shape every failure has on the wire.",
        "required": [
          "code",
          "message",
          "status"
        ],
        "properties": {
          "cause": {
            "type": [
              "string",
              "null"
            ],
            "description": "The chain of causes. It is absent if a user can see the response."
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "Description": {
        "type": "string",
        "maxLength": 1000,
        "minLength": 0
      },
...
      "GameMode": {
        "type": "string",
        "maxLength": 30,
        "minLength": 1
      },
      "GameName": {
        "type": "string",
        "maxLength": 40,
        "minLength": 1
      },
...
      "NewTournament": {
        "type": "object",
        "description": "Body of `POST /api/admin/tournaments`. A new tournament is always a draft.",
        "required": [
          "name",
          "game",
          "mode",
          "date",
          "registrationClosesAt"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "$ref": "#/components/schemas/Description"
          },
          "game": {
            "$ref": "#/components/schemas/GameName"
          },
          "mode": {
            "$ref": "#/components/schemas/GameMode"
          },
          "name": {
            "$ref": "#/components/schemas/TournamentName"
          },
          "registrationClosesAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "Paginated_Tournament": {
        "type": "object",
        "description": "One page of results, and the totals a client needs for a pager.",
        "required": [
          "items",
          "total",
          "page",
          "pageSize",
          "totalPages"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "A tournament as every client sees it. `entrant_count` and `has_bracket` ride\nalong so a list needs no second call per row.",
              "required": [
                "id",
                "name",
                "game",
                "mode",
                "description",
                "date",
                "registrationClosesAt",
                "status",
                "entrantCount",
                "hasBracket",
                "createdAt"
              ],
              "properties": {
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "date": {
                  "type": "string",
                  "format": "date",
                  "description": "The day of the event. A day has no timezone."
                },
                "description": {
                  "$ref": "#/components/schemas/Description"
                },
                "entrantCount": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                },
                "game": {
                  "$ref": "#/components/schemas/GameName"
                },
                "hasBracket": {
                  "type": "boolean"
                },
                "id": {
                  "$ref": "#/components/schemas/TournamentId"
                },
                "mode": {
                  "$ref": "#/components/schemas/GameMode"
                },
                "name": {
                  "$ref": "#/components/schemas/TournamentName"
                },
                "registrationClosesAt": {
                  "type": "string",
                  "format": "date-time",
                  "description": "Registrations close at this instant. Past it, nobody applies or retires."
                },
                "status": {
                  "$ref": "#/components/schemas/TournamentStatus"
                },
                "winner": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/Entrant"
                    }
                  ]
                }
              }
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "maximum": 100,
            "minimum": 1
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Every row that matches, ignoring the window.",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "RegistrationRequest": {
        "type": "object",
        "description": "Body of `POST /api/tournaments/{id}/registration`. A player applies with the\nlevel they give themself. A second apply with another level changes it.",
        "required": [
          "skill"
        ],
        "properties": {
          "skill": {
            "$ref": "#/components/schemas/SkillLevel"
          }
        },
        "additionalProperties": false
      },
      "Registrations": {
        "type": "object",
        "description": "Answer of `GET /api/me/registrations`: the tournaments the caller entered.\nOne call tells a page which apply buttons to turn into retire buttons.",
        "required": [
          "tournaments"
        ],
        "properties": {
          "tournaments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TournamentId"
            }
          }
        }
      },
...
      "Tournament": {
        "type": "object",
        "description": "A tournament as every client sees it. `entrant_count` and `has_bracket` ride\nalong so a list needs no second call per row.",
        "required": [
          "id",
          "name",
          "game",
          "mode",
          "description",
          "date",
          "registrationClosesAt",
          "status",
          "entrantCount",
          "hasBracket",
          "createdAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "The day of the event. A day has no timezone."
          },
          "description": {
            "$ref": "#/components/schemas/Description"
          },
          "entrantCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "game": {
            "$ref": "#/components/schemas/GameName"
          },
          "hasBracket": {
            "type": "boolean"
          },
          "id": {
            "$ref": "#/components/schemas/TournamentId"
          },
          "mode": {
            "$ref": "#/components/schemas/GameMode"
          },
          "name": {
            "$ref": "#/components/schemas/TournamentName"
          },
          "registrationClosesAt": {
            "type": "string",
            "format": "date-time",
            "description": "Registrations close at this instant. Past it, nobody applies or retires."
          },
          "status": {
            "$ref": "#/components/schemas/TournamentStatus"
          },
          "winner": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Entrant"
              }
            ]
          }
        }
      },
      "TournamentDetail": {
        "type": "object",
        "description": "One tournament with everything a page needs.",
        "required": [
          "tournament",
          "entrants"
        ],
        "properties": {
          "bracket": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Bracket"
              }
            ]
          },
          "entrants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Entrant"
            }
          },
          "tournament": {
            "$ref": "#/components/schemas/Tournament"
          }
        }
      },
      "TournamentId": {
        "type": "string",
        "format": "uuid"
      },
...
      "TournamentUpdate": {
        "type": "object",
        "description": "Body of `PATCH /api/admin/tournaments/{id}`. An absent field keeps its value.",
        "properties": {
          "date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Description"
              }
            ]
          },
          "game": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GameName"
              }
            ]
          },
          "mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GameMode"
              }
            ]
          },
          "name": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TournamentName"
              }
            ]
          },
          "registrationClosesAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}